refactor: change structure of the project and clearing code
