+++ js/generalFunctions.js
